LoadingPage.xaml MainPage.xaml
