+++ ApriSiSteam.WPF/Pages/LoadingPage.xaml
@@ -1,0 +1,14 @@
+﻿<Page x:Class="ApriSiSteam.WPF.Pages.LoadingPage"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:ApriSiSteam.WPF.Pages"
+      mc:Ignorable="d" 
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="LoadingPage">
+
+    <Grid>
+        <TextBlock Foreground="White" FontSize="150">Hello World!</TextBlock>
+    </Grid>
+</Page>
